Added Worker state machine.
State machine that kicks off worker and waits till it's done.

Currently uses LeaseHolders to monitor processes, but thinking it could be
refactored into this class.  That could cut out the need for
LeaseHolders and maybe LeaseHolderSupervisors.

A re-think of the supervision tree could simplify things a bit as well -
maybe just terminate the python etc. processes if they're running when
the workers go down.  Not the end of the world, provided we haven't
requested a ridiculous lease time on the tasks.