Refactored LeaseHolder etc. into Worker
Worker and supervisor were both doing really little, and it was just
complicating the code somewhat (particularly since we had a
LeaseHolderSupervisor as well), so I've moved the responsibilities of
LeaseHolder into Worker.

An added bonus is that we'll now avoid the sitation where leaseholder
monitors after job is done and thinks that job ended abnormally.