Added WorkerSupervisor
This supervisor is responsible for managing a pair of processes: the
python process and a worker that utilises it.  It uses an all_for_one
restart strategy - since errors on python will break the worker and
errors in the worker should kill any task running in python.

Hopefully the terminate handler in worker should also ensure that tasks
get released when either process goes down.