Created internal Elk.Task record.
Elk.Task plays nicely with Inspect protocol, and has it's own protocols
in GoogleAPI for translation to & from the GoogleAPI HashDict.

For the sake of simplicity I'm just storing the HashDict in the record
itself.  Means we don't have to make the record extensive, just extract
what we need, then just send the original data back to google.