Added support for gzipped data to GoogleAPIReader.
The app engine task queue seems to have a wierd bug.  In certain
circumstances it will send totally corrupted data to the client - data
that is missing padding, and only some of the existing decoded data
makes sense.  From my limited testing, it seems that this is triggered
by ?s in the data, since it's always ?s that start the corrupted data,
and removing them causes success.

Since I can't change appengine, and removing ?s from all data is
unrealistic, I'm adding a gzipped mode to Elk.  This mode will accept
gzipped, base64 encoded data in the "payload" parameter.  Base64
encoding doesn't have ?s, so we should be safe with this mode.