Logging now less verbose when idle.
Currently Elk outputs 2 lines of info logs every 2 seconds.  This is
fine for debugging things, but a bit excessive for production, where it
could obscure useful information.  To counter this, I've switched both
the lines to debug level.  The "x tasks leased" message will still be an
info message, but only when tasks are actually leased.